Typos, structure
A few typos and similar, as well as a challenge that was introduced before its concepts had been. Also the end of a sentence was missing so I filled that in based on what I thought your point was, but maybe I was wrong.

And I'm not entirely sure what you mean by the following? Maybe worth expanding upon.

>Since a node is _either_ an atom (value) _or_ a cell (fork), you never have to decide if the contents of a node is a direct value or a tree:  it just happens.